refactor: move game code into service
