+++ src/pages/api/createGame.ts
@@ -1,25 +1,10 @@
 import { Request, Response } from 'express';
-import fetch from 'isomorphic-unfetch';
-const setId = 'JP8FV';
-const cardcastUrl = `https://api.cardcastgame.com/v1/decks/${setId}/cards`;
+import { getGame } from '../../services/gameService';
 
 export default async (_req: Request, res: Response): Promise<void> => {
     try {
-        const fetchResponse = await fetch(cardcastUrl);
-        const data: CardcastCardSet = await fetchResponse.json();
-
-        const blackCards = data.calls.map((card: CardcastCard) => {
-            const text = card.text.join('____');
-            return { text };
-        });
-        const whiteCards = data.responses.map((card: CardcastCard) => {
-            const text = card.text.join('');
-            return { text };
-        });
-        res.status(200).json({
-            blackCards,
-            whiteCards,
-        });
+        const game = await getGame();
+        res.status(200).json(game);
     } catch (err) {
         console.error(err);
         res.status(500).end();
